Review system: add multi-user review model
This commit adds the database migration to add a separate review table to the
database. Each entry represents a review by a particular user for a particular
treenode of a skeleton. This new table has the following fields: reviewer,
reviewer_time, treenode and skeleton. Technically, the last field (skeleton)
isn't needed, because one can get the skeleton through the treenode. However, it
makes sense to have it for performance reasons.

This is part of issue #580.